lease-server: Reuse unused lease socket names
Before starting a lease server, any sockets of the same name from
a previous run should be deleted so that the new instance
can bind to the address.

An advisory lock is used to check that the socket is not in use by
an already running instance of the server before deleting it.

Signed-off-by: Damian Hobson-Garcia <dhobsong@igel.co.jp>