lease-server: Allow multiple client connections
Allow multiple clients to issue lease requests
on a server at the same time. This is
necessary to be able to grant or deny leases,
not just on a first-come-first-served basis.

Future patches will add extra contitions,
such as command-line options and lease configuration
settings to determine when and how lease requests should
be granted.

This update changes the behaviour of the lease-server interface
so that it reports every client connection request, instead of
when a server has accepted a request, so update the test suite to
reflect this.

Bug-AGL: SPEC-3816

Signed-off-by: Damian Hobson-Garcia <dhobsong@igel.co.jp>
Change-Id: I08506b6ebc2a4910e7553e3dfd9b1e230ea0008c