Add option to hold lease on unexpected disconnect
If a client application crashes while owning a lease,
the lease will automatically be revoked and all references
freed to make it available for another (or the restarted
original) client to claim and continue display.  This will
cause the screen to blank while the lease is unowned.

Add an option to revoke the lease, but keep the reference to
the lease descriptor, so that the screen does not blank while
unowned.

This only has effect when a client exits without calling the
`dlm_release_lease` function. Otherwise the lease is revoked
and the descriptor dereferenced, just as before.

Bug-AGL: SPEC-3862

Signed-off-by: Damian Hobson-Garcia <dhobsong@igel.co.jp>
Change-Id: I3fc2c8fde4118b6e298ba31807fd5fa4766a97dd